+++ package.json
@@ -56,6 +56,7 @@
     "cm6-theme-basic-dark": "^0.2.0",
     "cm6-theme-basic-light": "^0.2.0",
     "codemirror": "^6.0.1",
+    "codemirror-lang-mermaid": "^0.5.0",
     "diff": "^8.0.3",
     "highlight.js": "^11.9.0",
     "jsonpath": "^1.2.1",
